new test file added, DynamicMemberLookupTests.swift
